# tests/snapshots/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api-starter-axum",
    "description": "REST API starter: users, auth (password, OAuth, passkeys, QR-code, TOTP), API keys, audit log.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "get_one",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed to view this user"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "users"
        ],
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User soft-deleted"
          },
          "400": {
            "description": "Cannot delete self, or would remove the last admin"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "users"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Would lock the actor out or remove the last admin"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Role": {
        "type": "string",
        "enum": [
          "user",
          "admin"
        ]
      },
...
      "UpdateUserDto": {
        "type": "object",
        "description": "Admin PATCH: every field is optional, absent means \"leave unchanged\".",
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "role": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "is_active",
          "email_verified",
          "has_password",
          "totp_enabled",
          "created_at"
        ],
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Relative URL of the profile photo (public, cacheable), or null."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "has_password": {
            "type": "boolean",
            "description": "`false` for accounts created through OAuth that never chose a password."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_active": {
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "totp_enabled": {
            "type": "boolean",
            "description": "Whether two-factor authentication (TOTP) is turned on for this account."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A session JWT (from /auth/login or /auth/2fa/verify) or an API key (ak_...)"
      }
    }
  },
...
}